Refactor tree generation
