+++ CHANGELOG.md
@@ -1,11 +1,7 @@
 # Changelog
 
-## 2025-07-26
-- 1042 Split assets.json into external category lists and update loader
-- 1008 Move assets.json and mats.json into json directory and update paths.
-- 0930 Fix all ground texture paths to use local assets.
-- 0915 Fix all material and ground textures to load from local relative paths instead of external URLs.
-- 0906 Fixed ground textures loading from external URLs causing transparent terrain
+## 2025-07-29
+- 0953 Break up trees.js into separate tree modules
 
 ## Guidelines for future updates
 - List changes in reverse chronological order (newest first).
